requirements.txt
pydantic
typing-extensions
prefect
requests
python-dotenv
langchain-core
langchain-community
beautifulsoup4